+++ milkman/src/main/java/milkman/ui/commands/UiCommand.java
@@ -35,6 +35,9 @@ public interface UiCommand {
 	public static class SubmitActiveRequest implements UiCommand {
 	}
 	@Value
+	public static class CancelActiveRequest implements UiCommand {
+	}
+	@Value
 	public static class LoadRequest implements UiCommand {
 		String requestId;
 	}